requirements.txt
google-generativeai
pyautogui
python-dotenv
pynput
Pillow
matplotlib
openai